# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de autenticación",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        },
        {
            "url": "https://liryca-api.vercel.app/api"
        }
    ],
    "paths": {
...
        "/api/v1/song/top-by-artist/{artistId}": {
            "get": {
                "tags": [
                    "song"
                ],
                "summary": "Get top songs by artist",
                "description": "Returns the top 10 songs by an artist based on playback count",
                "operationId": "top-by-artist",
                "parameters": [
                    {
                        "name": "artistId",
                        "in": "path",
                        "description": "ID of the artist",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Mírame"
                                                    },
                                                    "artist_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "album_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "time": {
                                                        "type": "string",
                                                        "example": "00:01:38"
                                                    },
                                                    "genre": {
                                                        "type": "string",
                                                        "example": "Reggaeton"
                                                    },
                                                    "url_song": {
                                                        "type": "string",
                                                        "example": "https://i.scdn.co/image/ab67616d0000b273b62a2ec2d61d48f34a368144"
                                                    },
                                                    "is_active": {
                                                        "type": "boolean",
                                                        "example": true
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-11-18T00:20:29.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-11-18T00:20:29.000000Z"
                                                    },
                                                    "play_count": {
                                                        "type": "integer",
                                                        "example": 1
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Error: Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "No songs found for this artist"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/song/by-album-id/{albumId}": {
            "get": {
                "tags": [
                    "song"
                ],
                "summary": "Get songs by album",
                "description": "Returns all songs from a specific album",
                "operationId": "by-album-id",
                "parameters": [
                    {
                        "name": "albumId",
                        "in": "path",
                        "description": "ID of the album",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Mírame"
                                                    },
                                                    "artist_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "album_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "time": {
                                                        "type": "string",
                                                        "example": "00:01:38"
                                                    },
                                                    "genre": {
                                                        "type": "string",
                                                        "example": "Reggaeton"
                                                    },
                                                    "url_song": {
                                                        "type": "string",
                                                        "example": "https://i.scdn.co/image/ab67616d0000b273b62a2ec2d61d48f34a368144"
                                                    },
                                                    "is_active": {
                                                        "type": "boolean",
                                                        "example": true
                                                    },
                                                    "album_icon": {
                                                        "type": "string",
                                                        "example": "https://i.scdn.co/image/ab67616d0000b273b62a2ec2d61d48f34a368144"
                                                    },
                                                    "artist_username": {
                                                        "type": "string",
                                                        "example": "artist_username"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-11-18T00:20:29.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-11-18T00:20:29.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Error: Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "No songs found for this album"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "sanctum"
            }
        }
    }
}